refactor parsers in `input_validators` to `input_parsers`
