+++ clap-v3-utils/src/input_validators.rs
@@ -181,30 +181,6 @@ where
     }
 }
 
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub struct PubkeySignature {
-    pubkey: Pubkey,
-    signature: Signature,
-}
-impl FromStr for PubkeySignature {
-    type Err = String;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let mut signer = s.split('=');
-        let pubkey = signer
-            .next()
-            .ok_or_else(|| String::from("Malformed signer string"))?;
-        let pubkey = Pubkey::from_str(pubkey).map_err(|err| format!("{err}"))?;
-
-        let signature = signer
-            .next()
-            .ok_or_else(|| String::from("Malformed signer string"))?;
-        let signature = Signature::from_str(signature).map_err(|err| format!("{err}"))?;
-
-        Ok(Self { pubkey, signature })
-    }
-}
-
 // Return an error if a url cannot be parsed.
 #[deprecated(since = "1.17.0", note = "please use `parse_url` instead")]
 pub fn is_url<T>(string: T) -> Result<(), String>
@@ -223,19 +199,6 @@ where
     }
 }
 
-pub fn parse_url(arg: &str) -> Result<String, String> {
-    match url::Url::parse(arg) {
-        Ok(url) => {
-            if url.has_host() {
-                Ok(arg.to_string())
-            } else {
-                Err("no host provided".to_string())
-            }
-        }
-        Err(err) => Err(format!("{err}")),
-    }
-}
-
 #[deprecated(since = "1.17.0", note = "please use `parse_url_or_moniker` instead")]
 pub fn is_url_or_moniker<T>(string: T) -> Result<(), String>
 where
@@ -245,19 +208,6 @@ where
         Ok(url) => {
             if url.has_host() {
                 Ok(())
-            } else {
-                Err("no host provided".to_string())
-            }
-        }
-        Err(err) => Err(format!("{err}")),
-    }
-}
-
-pub fn parse_url_or_moniker(arg: &str) -> Result<String, String> {
-    match url::Url::parse(&normalize_to_url_if_moniker(arg)) {
-        Ok(url) => {
-            if url.has_host() {
-                Ok(arg.to_string())
             } else {
                 Err("no host provided".to_string())
             }
@@ -316,18 +266,6 @@ where
         })
 }
 
-pub fn parse_pow2(arg: &str) -> Result<usize, String> {
-    arg.parse::<usize>()
-        .map_err(|e| format!("Unable to parse, provided: {arg}, err: {e}"))
-        .and_then(|v| {
-            if !v.is_power_of_two() {
-                Err(format!("Must be a power of 2: {v}"))
-            } else {
-                Ok(v)
-            }
-        })
-}
-
 #[deprecated(
     since = "1.17.0",
     note = "please use `clap_value_parser!(u16)` instead"
@@ -359,20 +297,6 @@ where
         })
 }
 
-pub fn parse_percentage(arg: &str) -> Result<u8, String> {
-    arg.parse::<u8>()
-        .map_err(|e| format!("Unable to parse input percentage, provided: {arg}, err: {e}"))
-        .and_then(|v| {
-            if v > 100 {
-                Err(format!(
-                    "Percentage must be in range of 0 to 100, provided: {v}"
-                ))
-            } else {
-                Ok(v)
-            }
-        })
-}
-
 #[deprecated(
     since = "1.17.0",
     note = "please use `UiTokenAmount::parse_amount` instead"
@@ -388,53 +312,6 @@ where
             "Unable to parse input amount as integer or float, provided: {amount}"
         ))
     }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub enum UiTokenAmount {
-    Amount(f64),
-    All,
-}
-impl UiTokenAmount {
-    pub fn parse_amount(arg: &str) -> Result<UiTokenAmount, String> {
-        if let Ok(amount) = arg.parse::<f64>() {
-            Ok(UiTokenAmount::Amount(amount))
-        } else {
-            Err(format!("Unable to parse input amount, provided: {arg}"))
-        }
-    }
-
-    pub fn parse_amount_or_all(arg: &str) -> Result<UiTokenAmount, String> {
-        if let Ok(amount) = arg.parse::<f64>() {
-            Ok(UiTokenAmount::Amount(amount))
-        } else if arg == "ALL" {
-            Ok(UiTokenAmount::All)
-        } else {
-            Err(format!(
-                "Unable to parse input amount as float or 'ALL' keyword, provided: {arg}"
-            ))
-        }
-    }
-
-    pub fn to_raw_amount(&self, decimals: u8) -> RawTokenAmount {
-        match self {
-            UiTokenAmount::Amount(amount) => {
-                RawTokenAmount::Amount((amount * 10_usize.pow(decimals as u32) as f64) as u64)
-            }
-            UiTokenAmount::All => RawTokenAmount::All,
-        }
-    }
-
-    pub fn sol_to_lamport(&self) -> RawTokenAmount {
-        const NATIVE_SOL_DECIMALS: u8 = 9;
-        self.to_raw_amount(NATIVE_SOL_DECIMALS)
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub enum RawTokenAmount {
-    Amount(u64),
-    All,
 }
 
 #[deprecated(
@@ -467,12 +344,6 @@ where
         .map_err(|e| format!("{e}"))
 }
 
-pub fn parse_rfc3339_datetime(arg: &str) -> Result<String, String> {
-    DateTime::parse_from_rfc3339(arg)
-        .map(|_| arg.to_string())
-        .map_err(|e| format!("{e}"))
-}
-
 #[deprecated(since = "1.17.0", note = "please use `parse_derivation` instead")]
 pub fn is_derivation<T>(value: T) -> Result<(), String>
 where
@@ -494,25 +365,6 @@ where
             }
         })
         .map(|_| ())
-}
-
-pub fn parse_derivation(arg: &str) -> Result<String, String> {
-    let value = arg.replace('\'', "");
-    let mut parts = value.split('/');
-    let account = parts.next().unwrap();
-    account
-        .parse::<u32>()
-        .map_err(|e| format!("Unable to parse derivation, provided: {account}, err: {e}"))
-        .and_then(|_| {
-            if let Some(change) = parts.next() {
-                change.parse::<u32>().map_err(|e| {
-                    format!("Unable to parse derivation, provided: {change}, err: {e}")
-                })
-            } else {
-                Ok(0)
-            }
-        })?;
-    Ok(arg.to_string())
 }
 
 #[deprecated(since = "1.17.0", note = "please use `parse_structured_seed` instead")]
@@ -561,47 +413,6 @@ where
     }
 }
 
-pub fn parse_structured_seed(arg: &str) -> Result<String, String> {
-    let (prefix, value) = arg
-        .split_once(':')
-        .ok_or("Seed must contain ':' as delimiter")
-        .unwrap();
-    if prefix.is_empty() || value.is_empty() {
-        Err(String::from("Seed prefix or value is empty"))
-    } else {
-        match prefix {
-            "string" | "pubkey" | "hex" | "u8" => Ok(arg.to_string()),
-            _ => {
-                let len = prefix.len();
-                if len != 5 && len != 6 {
-                    Err(format!("Wrong prefix length {len} {prefix}:{value}"))
-                } else {
-                    let sign = &prefix[0..1];
-                    let type_size = &prefix[1..len.saturating_sub(2)];
-                    let byte_order = &prefix[len.saturating_sub(2)..len];
-                    if sign != "u" && sign != "i" {
-                        Err(format!("Wrong prefix sign {sign} {prefix}:{value}"))
-                    } else if type_size != "16"
-                        && type_size != "32"
-                        && type_size != "64"
-                        && type_size != "128"
-                    {
-                        Err(format!(
-                            "Wrong prefix type size {type_size} {prefix}:{value}"
-                        ))
-                    } else if byte_order != "le" && byte_order != "be" {
-                        Err(format!(
-                            "Wrong prefix byte order {byte_order} {prefix}:{value}"
-                        ))
-                    } else {
-                        Ok(arg.to_string())
-                    }
-                }
-            }
-        }
-    }
-}
-
 #[deprecated(
     since = "1.17.0",
     note = "please use `parse_derived_address_seed` instead"
@@ -617,263 +428,5 @@ where
         ))
     } else {
         Ok(())
-    }
-}
-
-pub fn parse_derived_address_seed(arg: &str) -> Result<String, String> {
-    if arg.len() > MAX_SEED_LEN {
-        Err(format!(
-            "Address seed must not be longer than {MAX_SEED_LEN} bytes"
-        ))
-    } else {
-        Ok(arg.to_string())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use {
-        super::*,
-        clap::{Arg, Command},
-    };
-
-    #[test]
-    fn test_parse_pubkey() {
-        let command = Command::new("test").arg(
-            Arg::new("pubkey")
-                .long("pubkey")
-                .takes_value(true)
-                .value_parser(clap::value_parser!(Pubkey)),
-        );
-
-        // success case
-        let matches = command
-            .clone()
-            .try_get_matches_from(vec!["test", "--pubkey", "11111111111111111111111111111111"])
-            .unwrap();
-        assert_eq!(
-            *matches.get_one::<Pubkey>("pubkey").unwrap(),
-            Pubkey::from_str("11111111111111111111111111111111").unwrap(),
-        );
-
-        // validation fails
-        let matches_error = command
-            .clone()
-            .try_get_matches_from(vec!["test", "--pubkey", "this_is_an_invalid_arg"])
-            .unwrap_err();
-        assert_eq!(matches_error.kind, clap::error::ErrorKind::ValueValidation);
-    }
-
-    #[test]
-    fn test_parse_hash() {
-        let command = Command::new("test").arg(
-            Arg::new("hash")
-                .long("hash")
-                .takes_value(true)
-                .value_parser(clap::value_parser!(Hash)),
-        );
-
-        // success case
-        let matches = command
-            .clone()
-            .try_get_matches_from(vec!["test", "--hash", "11111111111111111111111111111111"])
-            .unwrap();
-        assert_eq!(
-            *matches.get_one::<Hash>("hash").unwrap(),
-            Hash::from_str("11111111111111111111111111111111").unwrap(),
-        );
-
-        // validation fails
-        let matches_error = command
-            .clone()
-            .try_get_matches_from(vec!["test", "--hash", "this_is_an_invalid_arg"])
-            .unwrap_err();
-        assert_eq!(matches_error.kind, clap::error::ErrorKind::ValueValidation);
-    }
-
-    #[test]
-    fn test_parse_pubkey_signature() {
-        let command = Command::new("test").arg(
-            Arg::new("pubkeysig")
-                .long("pubkeysig")
-                .takes_value(true)
-                .value_parser(clap::value_parser!(PubkeySignature)),
-        );
-
-        // success case
-        let matches = command
-            .clone()
-            .try_get_matches_from(vec![
-                "test", 
-                "--pubkeysig", 
-                "11111111111111111111111111111111=4TpFuec1u4BZfxgHg2VQXwvBHANZuNSJHmgrU34GViLAM5uYZ8t7uuhWMHN4k9r41B2p9mwnHjPGwTmTxyvCZw63"
-                ]
-            )
-            .unwrap();
-
-        let expected = PubkeySignature {
-            pubkey: Pubkey::from_str("11111111111111111111111111111111").unwrap(),
-            signature: Signature::from_str("4TpFuec1u4BZfxgHg2VQXwvBHANZuNSJHmgrU34GViLAM5uYZ8t7uuhWMHN4k9r41B2p9mwnHjPGwTmTxyvCZw63").unwrap(),
-        };
-
-        assert_eq!(
-            *matches.get_one::<PubkeySignature>("pubkeysig").unwrap(),
-            expected,
-        );
-
-        // validation fails
-        let matches_error = command
-            .clone()
-            .try_get_matches_from(vec!["test", "--pubkeysig", "this_is_an_invalid_arg"])
-            .unwrap_err();
-        assert_eq!(matches_error.kind, clap::error::ErrorKind::ValueValidation);
-    }
-
-    #[test]
-    fn test_parse_token_amount() {
-        let command = Command::new("test").arg(
-            Arg::new("amount")
-                .long("amount")
-                .takes_value(true)
-                .value_parser(UiTokenAmount::parse_amount),
-        );
-
-        // success cases
-        let matches = command
-            .clone()
-            .try_get_matches_from(vec!["test", "--amount", "11223344"])
-            .unwrap();
-        assert_eq!(
-            *matches.get_one::<UiTokenAmount>("amount").unwrap(),
-            UiTokenAmount::Amount(11223344_f64),
-        );
-
-        let matches = command
-            .clone()
-            .try_get_matches_from(vec!["test", "--amount", "0.11223344"])
-            .unwrap();
-        assert_eq!(
-            *matches.get_one::<UiTokenAmount>("amount").unwrap(),
-            UiTokenAmount::Amount(0.11223344),
-        );
-
-        // validation fail cases
-        let matches_error = command
-            .clone()
-            .try_get_matches_from(vec!["test", "--amount", "this_is_an_invalid_arg"])
-            .unwrap_err();
-        assert_eq!(matches_error.kind, clap::error::ErrorKind::ValueValidation);
-
-        let matches_error = command
-            .clone()
-            .try_get_matches_from(vec!["test", "--amount", "all"])
-            .unwrap_err();
-        assert_eq!(matches_error.kind, clap::error::ErrorKind::ValueValidation);
-    }
-
-    #[test]
-    fn test_parse_token_amount_or_all() {
-        let command = Command::new("test").arg(
-            Arg::new("amount")
-                .long("amount")
-                .takes_value(true)
-                .value_parser(UiTokenAmount::parse_amount_or_all),
-        );
-
-        // success cases
-        let matches = command
-            .clone()
-            .try_get_matches_from(vec!["test", "--amount", "11223344"])
-            .unwrap();
-        assert_eq!(
-            *matches.get_one::<UiTokenAmount>("amount").unwrap(),
-            UiTokenAmount::Amount(11223344_f64),
-        );
-
-        let matches = command
-            .clone()
-            .try_get_matches_from(vec!["test", "--amount", "0.11223344"])
-            .unwrap();
-        assert_eq!(
-            *matches.get_one::<UiTokenAmount>("amount").unwrap(),
-            UiTokenAmount::Amount(0.11223344),
-        );
-
-        let matches = command
-            .clone()
-            .try_get_matches_from(vec!["test", "--amount", "ALL"])
-            .unwrap();
-        assert_eq!(
-            *matches.get_one::<UiTokenAmount>("amount").unwrap(),
-            UiTokenAmount::All,
-        );
-
-        // validation fail cases
-        let matches_error = command
-            .clone()
-            .try_get_matches_from(vec!["test", "--amount", "this_is_an_invalid_arg"])
-            .unwrap_err();
-        assert_eq!(matches_error.kind, clap::error::ErrorKind::ValueValidation);
-    }
-
-    #[test]
-    fn test_sol_to_lamports() {
-        let command = Command::new("test").arg(
-            Arg::new("amount")
-                .long("amount")
-                .takes_value(true)
-                .value_parser(UiTokenAmount::parse_amount_or_all),
-        );
-
-        let test_cases = vec![
-            ("50", 50_000_000_000),
-            ("1.5", 1_500_000_000),
-            ("0.03", 30_000_000),
-        ];
-
-        for (arg, expected_lamport) in test_cases {
-            let matches = command
-                .clone()
-                .try_get_matches_from(vec!["test", "--amount", arg])
-                .unwrap();
-            assert_eq!(
-                matches
-                    .get_one::<UiTokenAmount>("amount")
-                    .unwrap()
-                    .sol_to_lamport(),
-                RawTokenAmount::Amount(expected_lamport),
-            );
-        }
-    }
-
-    #[test]
-    fn test_derivation() {
-        let command = Command::new("test").arg(
-            Arg::new("derivation")
-                .long("derivation")
-                .takes_value(true)
-                .value_parser(parse_derivation),
-        );
-
-        let test_arguments = vec![
-            ("2", true),
-            ("0", true),
-            ("65537", true),
-            ("0/2", true),
-            ("a", false),
-            ("4294967296", false),
-            ("a/b", false),
-            ("0/4294967296", false),
-        ];
-
-        for (arg, should_accept) in test_arguments {
-            if should_accept {
-                let matches = command
-                    .clone()
-                    .try_get_matches_from(vec!["test", "--derivation", arg])
-                    .unwrap();
-                assert_eq!(matches.get_one::<String>("derivation").unwrap(), arg);
-            }
-        }
     }
 }
