+++ clap-v3-utils/src/offline.rs
@@ -1,5 +1,5 @@
 use {
-    crate::{input_validators::*, ArgConstant},
+    crate::{input_parsers::PubkeySignature, ArgConstant},
     clap::{value_parser, Arg, Command},
     solana_sdk::hash::Hash,
 };
